+++ index.html
@@ -6,5 +6,6 @@
   </head>
   <body>
     <div id="root"></div>
+    <script src="dist/bundle.js"></script>
   </body>
 </html>
