+++ server/helperFunctions/scraper.js
